refactor: 검증부분 UserNames 클래스에 이전
